Add files via upload
Here are files containing the make-up of my personal-portfolio /Developer Cv website I created using Visual Studio Code. It was a challenging yet exciting project
since it is the first time I was ever able to create a website from scratch. While it might not be at its best, I am happy to have successfully finish it.
cheers to more future projects and the strength to finish well......